Move WebHostExtensions to own file.
